Added Bot ID and call to command file
When the bot logs in the bot's ID will print in console. This will help with testing. I removed '!' from commands and set all letter to lower case. The command will then be search in the command folder and run the appropriate file. If the file is not found an error message will print.